сберспасибо

+++ src/main/java/ru/sbrf/hackaton/telegram/bot/model/GeoPosition.java
@@ -14,9 +14,9 @@ public class GeoPosition {
     private Long id;
 
     @Column(name="longitude", nullable = false)
-    private Double longitude;
+    private Float longitude;
     @Column(name="latitude", nullable = false)
-    private Double latitude;
+    private Float latitude;
 
     public Long getId() {
         return id;
@@ -26,19 +26,19 @@ public class GeoPosition {
         this.id = id;
     }
 
-    public Double getLongitude() {
+    public Float getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Double longitude) {
+    public void setLongitude(Float longitude) {
         this.longitude = longitude;
     }
 
-    public Double getLatitude() {
+    public Float getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Double latitude) {
+    public void setLatitude(Float latitude) {
         this.latitude = latitude;
     }
